requirements.txt
streamlit
requests
python-dotenv
SpeechRecognition
deep-translator
googletrans



